+++ packages/server/api/src/app/ai/mcp/openops-tools.ts
@@ -16,7 +16,6 @@ import { MCPTool } from './mcp-tools';
 const INCLUDED_PATHS: Record<string, string[]> = {
   '/v1/files/{fileId}': ['get'],
   '/v1/flow-versions/': ['get'],
-  '/v1/flow-versions/{flowVersionId}/{stepId}/test': ['post'],
   '/v1/flows/': ['get'],
   '/v1/flows/count': ['get'],
   '/v1/flows/{id}': ['get'],
@@ -30,6 +29,7 @@ const INCLUDED_PATHS: Record<string, string[]> = {
   '/v1/app-connections/': ['get', 'patch'],
   '/v1/app-connections/{id}': ['get'],
   '/v1/app-connections/metadata': ['get'],
+  '/v1/test/{flowVersionId}/{stepId}': ['post'],
 };
 
 function filterOpenApiSchema(schema: OpenAPI.Document): OpenAPI.Document {
