+++ packages/server/api/src/app/flows/flow.module.ts
@@ -25,11 +25,13 @@ import { flowVersionController } from './flow/flow-version.controller';
 import { flowController } from './flow/flow.controller';
 import { stepRunService } from './step-run/step-run-service';
 import { testTriggerController } from './test-trigger/test-trigger-controller';
+import { testController } from './test/test.controller';
 
 export const flowModule: FastifyPluginAsyncTypebox = async (app) => {
   await app.register(flowWorkerController, { prefix: '/v1/worker/flows' });
   await app.register(flowVersionController, { prefix: '/v1/flow-versions' });
   await app.register(flowController, { prefix: '/v1/flows' });
+  await app.register(testController, { prefix: '/v1/test' });
   await app.register(testTriggerController, { prefix: '/v1/test-trigger' });
   websocketService.addListener(WebsocketServerEvent.TEST_FLOW_RUN, (socket) => {
     return async (data: TestFlowRunRequestBody) => {
