+++ packages/server/api/src/app/flows/flow/flow-version.controller.ts
@@ -11,12 +11,10 @@ import {
   SERVICE_KEY_SECURITY_OPENAPI,
   StepOutputWithData,
   UpdateFlowVersionRequest,
-  flowHelper,
 } from '@openops/shared';
 import { Type as T } from '@sinclair/typebox';
 import { StatusCodes } from 'http-status-codes';
 import { flowVersionService } from '../flow-version/flow-version.service';
-import { stepRunService } from '../step-run/step-run-service';
 import { flowStepTestOutputService } from '../step-test-output/flow-step-test-output.service';
 import { flowService } from './flow.service';
 
@@ -191,52 +189,6 @@ export const flowVersionController: FastifyPluginAsyncTypebox = async (
       });
     },
   );
-
-  fastify.post(
-    '/:flowVersionId/:stepId/test',
-    TestStepRequest,
-    async (request, reply) => {
-      const { flowVersionId, stepId } = request.params;
-      const projectId = request.principal.projectId;
-
-      const flowVersion = await flowVersionService.getOneOrThrow(flowVersionId);
-
-      const step = flowHelper
-        .getAllSteps(flowVersion.trigger)
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        .find((step: any) => step.id === stepId);
-
-      if (!step) {
-        await reply.status(StatusCodes.NOT_FOUND).send({
-          success: false,
-          output: 'Step not found',
-        });
-        return;
-      }
-
-      try {
-        const result = await stepRunService.create({
-          userId: request.principal.id,
-          projectId,
-          flowVersionId: flowVersion.id,
-          stepName: step.name,
-        });
-
-        await reply.send({
-          success: result.success,
-          output: result.output ?? {},
-        });
-      } catch (error) {
-        await reply.status(StatusCodes.BAD_REQUEST).send({
-          success: false,
-          output:
-            error instanceof Error
-              ? error.message
-              : 'An error occurred while testing the step',
-        });
-      }
-    },
-  );
 };
 
 const GetLatestVersionsByConnectionRequestOptions = {
@@ -288,31 +240,6 @@ const SaveTestOutputRequestOptions = {
       [StatusCodes.NOT_FOUND]: T.Object({
         success: T.Boolean(),
         message: T.String(),
-      }),
-    },
-  },
-};
-
-const TestStepRequest = {
-  schema: {
-    description:
-      'Test a flow step with specified parameters. With this endpoint its possible to validate steps.',
-    params: Type.Object({
-      flowVersionId: Type.String(),
-      stepId: Type.String(),
-    }),
-    response: {
-      [StatusCodes.OK]: Type.Object({
-        success: Type.Boolean(),
-        output: Type.Any(),
-      }),
-      [StatusCodes.BAD_REQUEST]: Type.Object({
-        success: Type.Boolean(),
-        output: Type.String(),
-      }),
-      [StatusCodes.NOT_FOUND]: Type.Object({
-        success: Type.Boolean(),
-        output: Type.String(),
       }),
     },
   },
